Making the database and making the app work with password protection.
The changes done to the code were. Making a functional databas to store the name username and password for the users of the app. Then I made a new function so the there was a posibility of making a account with their name, a username and a password.
For the app to work with the database two functions were made the first to insert the data of the name, the username and the password into the database, and the second to extract said data. To add a higher level of security the passwords are hashed before entering the database.
Finally I changed the Log_in function to be able to work with the data in the database.