Creating a table in the database for the countries.
the file my_app_functions.py was changed so that it containes a list of countries for the app to work. it will also have the chance to add a country if need be.
also in the file databese interace the functions to edit the database for the taxation where added.
aswell as changing the names of many of the files to maintain order.