extending the database.
Adding a new table to the database called brackets where all the tax brackets will be stored. Appart from this making a class to better control the variables in the database and mapping them to the database.
Also extending the mapping of the other classes to be able to update or delete rows if need be.